Structure in different files for better development
