+++ lib/collision/collision_sat.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module CollisionSAT
+  def self.is_overlaps?(a, b)
+    raise NotImplementedError.new
+  end
+end
