+++ lib/collision/collisions.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Collisions
+  def self.is_overlaps?(a, b)
+    raise NotImplementedError.new
+  end
+end
